requirements.txt
pandas==1.3.4
python_dateutil==2.8.2
pytz==2021.3
requests==2.26.0
